moved around some stuff and updated the example
