+++ examples/testgputext/shaders/shader.frag
