+++ examples/testgputext/shaders/shader.frag.spv
