+++ examples/testgputext/shaders/shader.vert.spv
